requirements.txt
transitions==0.6.9
paho-mqtt==1.4.0
gpiozero==1.5.0
